refactor(config): split discovery module (#22145)
Splits out the config discovery from the main `lib.rs` file.

`lib.rs` will now just re-export items.

See upstack PRs for more context.
